requirements.txt
fastapi[all]
requests
python-dotenv
pymupdf
pinecone
sentence-transformers 
einops
google-genai